new file cmd_path_utils for norm
