+++ src/parsing/cmd_path.c
@@ -6,43 +6,11 @@
 /*   By: csweetin <csweetin@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/10 17:25:53 by csweetin          #+#    #+#             */
-/*   Updated: 2024/04/23 17:37:06 by csweetin         ###   ########.fr       */
+/*   Updated: 2024/04/23 17:42:16 by csweetin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "parsing.h"
-
-int	path_len(char *env, int *i)
-{
-	int	len;
-
-	len = 0;
-	while (env[*i] && env[*i] != ':')
-	{
-		*i += 1;
-		len++;
-	}
-	return (len);
-}
-
-int	check_path(char **path)
-{
-	struct stat	buf;
-
-	if (access(*path, X_OK) == 0)
-	{
-		if (stat(*path, &buf) == -1)
-		{
-			free(*path);
-			*path = NULL;
-			return (1);
-		}
-		if (S_ISDIR(buf.st_mode))
-			return (0);
-		return (1);
-	}
-	return (0);
-}
 
 char	*find_path(char *cmd, char *env)
 {
